requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.8
Pillow==9.5.0
drf-spectacular==0.26.2
gunicorn==20.1.0
